last updates

+++ algebra/NArray.cc
@@ -8,5 +8,5 @@ NArray::NArray(std::vector<int> N){
     
     for(int i = 0; i < N.size(); i++)
         size *= N[i];
-    data = new int[size];
+    data = new float[size];
 }
